@@ -45,13 +45,13 @@ setup(
         "eia 861",
         "ferc 714",
     ],
-    python_requires=">=3.8,<3.11",
+    python_requires=">=3.10,<3.11",
     setup_requires=["setuptools_scm"],
     install_requires=[
         "addfips>=0.3.1,<0.4.0",
         "catalystcoop.dbfread>=3.0,<3.1",
         "coloredlogs>=15.0,<15.1",
-        "dask>=2021.8,<2022.5.3",
+        "dask>=2021.8,<2022.6.1",
         "datapackage>=1.11,<1.16",  # Transition datastore to use frictionless.
         # "email-validator>=1.0.3",  # pydantic[email] dependency
         "catalystcoop.ferc_xbrl_extractor @ git+https://github.com/catalyst-cooperative/ferc-xbrl-extractor.git@dev",
@@ -70,7 +70,7 @@ setup(
         "pyyaml>=5,<6.1",
         "scikit-learn>=1.0,<1.2",
         "scipy>=1.6,<1.9",
-        "sqlalchemy>=1.4,<1.4.37",
+        "sqlalchemy>=1.4,<1.4.38",
         "timezonefinder>=5,<6.1",
         "xlsxwriter>=3,<3.1",
     ],
@@ -80,18 +80,17 @@ setup(
             "ipdb>=0.13,<0.14",
             "isort>=5.0,<5.11",
             "jedi>=0.18,<0.19",
-            "lxml>=4.6,<4.9",
+            "lxml>=4.6,<4.10",
             "tox>=3.20,<4.0.0",
             "twine>=3.3,<4.1",
         ],
         "doc": [
             "doc8>=0.9,<0.12",
+            "furo>=2022.4.7",
             "sphinx>=4,<6.0",
             "sphinx-autoapi>=1.8,<1.9",
             "sphinx-issues>=1.2,<3.1",
             "sphinx-reredirects",
-            "sphinx-rtd-dark-mode>=1.2,<1.3",
-            "sphinx-rtd-theme>=1.0,<1.1",
             "sphinxcontrib_bibtex>=2.4,<2.5",
         ],
         "test": [
@@ -106,14 +105,14 @@ setup(
             "flake8-use-fstring>=1.0,<1.4",
             "mccabe>=0.6,<0.8",
             "nbval>=0.9,<0.10",
-            "pep8-naming>=0.12,<0.13",
+            "pep8-naming>=0.12,<0.14",
             "pre-commit>=2.9,<2.20",
             "pydocstyle>=5.1,<6.2",
             "pytest>=6.2,<7.2",
             "pytest-console-scripts>=1.1,<1.4",
             "pytest-cov>=2.10,<3.1",
             "responses>=0.14,<0.22",
-            "rstcheck[sphinx]>=5.0,<5.1",
+            "rstcheck[sphinx]>=5.0,<6.1",
             "tox>=3.20,<3.26",
         ],
         "datasette": [
@@ -130,8 +129,6 @@ setup(
         "Programming Language :: Python",
         "Programming Language :: Python :: 3",
         "Programming Language :: Python :: 3 :: Only",
-        "Programming Language :: Python :: 3.8",
-        "Programming Language :: Python :: 3.9",
         "Programming Language :: Python :: 3.10",
         "Topic :: Scientific/Engineering",
     ],
