@@ -55,7 +55,6 @@ import json
 import time
 from datetime import datetime
 from pathlib import Path
-from typing import Dict, Set
 
 import dbfread
 import pandas as pd
@@ -155,7 +154,7 @@ def missing_respondents(reported, observed, identified):
     return records
 
 
-def observed_respondents(ferc1_engine: sa.engine.Engine) -> Set[int]:
+def observed_respondents(ferc1_engine: sa.engine.Engine) -> set[int]:
     """Compile the set of all observed respondent IDs found in the FERC 1 database.
 
     A significant number of FERC 1 respondent IDs appear in the data tables, but not
@@ -171,7 +170,7 @@ def observed_respondents(ferc1_engine: sa.engine.Engine) -> Set[int]:
 
     """
     f1_table_meta = pudl.output.pudltabl.get_table_meta(ferc1_engine)
-    observed = set([])
+    observed = set()
     for table in f1_table_meta.values():
         if "respondent_id" in table.columns:
             observed = observed.union(
@@ -192,8 +191,8 @@ class Ferc1Datastore:
     def __init__(self, datastore: Datastore):
         """Instantiate datastore wrapper for ferc1 resources."""
         self.datastore = datastore
-        self._cache: Dict[int, io.BytesIO] = {}
-        self.dbc_path: Dict[int, Path] = {}
+        self._cache: dict[int, io.BytesIO] = {}
+        self.dbc_path: dict[int, Path] = {}
 
         with importlib.resources.open_text(self.PACKAGE_PATH, "file_map.csv") as f:
             for row in csv.DictReader(f):
@@ -528,7 +527,7 @@ class FERC1FieldParser(dbfread.FieldParser):
         # Replace bare periods (which are non-numeric) with zero.
         if data == b".":
             data = b"0"
-        return super(FERC1FieldParser, self).parseN(field, data)
+        return super().parseN(field, data)
 
 
 def get_raw_df(
