@@ -16,7 +16,7 @@ from collections import defaultdict
 from functools import partial
 from importlib import resources
 from io import BytesIO
-from typing import Any, DefaultDict, Dict, List, Optional, Set, Union
+from typing import Any, Literal
 
 import addfips
 import coloredlogs
@@ -25,7 +25,7 @@ import pandas as pd
 import requests
 import sqlalchemy as sa
 
-from pudl.metadata.fields import apply_pudl_dtypes, get_pudl_dtypes
+from pudl.metadata.fields import get_pudl_dtypes
 
 sum_na = partial(pd.Series.sum, skipna=False)
 """A sum function that returns NA if the Series includes any NA values.
@@ -51,8 +51,8 @@ def label_map(
     df: pd.DataFrame,
     from_col: str = "code",
     to_col: str = "label",
-    null_value: Union[str, type(pd.NA)] = pd.NA,
-) -> DefaultDict[str, Union[str, type(pd.NA)]]:
+    null_value: str | type(pd.NA) = pd.NA,
+) -> defaultdict[str, str | type(pd.NA)]:
     """Build a mapping dictionary from two columns of a labeling / coding dataframe.
 
     These dataframes document the meanings of the codes that show up in much of the
@@ -84,9 +84,9 @@ def label_map(
 def find_new_ferc1_strings(
     table: str,
     field: str,
-    strdict: Dict[str, List[str]],
+    strdict: dict[str, list[str]],
     ferc1_engine: sa.engine.Engine,
-) -> Set[str]:
+) -> set[str]:
     """Identify as-of-yet uncategorized freeform strings in FERC Form 1.
 
     Args:
@@ -110,7 +110,7 @@ def find_new_ferc1_strings(
     return all_strings.difference(old_strings)
 
 
-def find_foreign_key_errors(dfs: Dict[str, pd.DataFrame]) -> List[Dict[str, Any]]:
+def find_foreign_key_errors(dfs: dict[str, pd.DataFrame]) -> list[dict[str, Any]]:
     """Report foreign key violations from a dictionary of dataframes.
 
     The database schema to check against is generated based on the names of the
@@ -342,90 +342,271 @@ def is_doi(doi):
     return bool(re.match(doi_regex, doi))
 
 
-def clean_merge_asof(
-    left: pd.DataFrame,
-    right: pd.DataFrame,
-    left_on: str = "report_date",
-    right_on: str = "report_date",
-    by: List[str] = [],
-) -> pd.DataFrame:
-    """Merge two dataframes having different ``report_date`` frequencies.
+def convert_col_to_datetime(df, date_col_name):
+    """Convert a column in a dataframe to a datetime.
 
-    We often need to bring together data which is reported on a monthly basis, and
-    entity attributes that are reported on an annual basis.  The
-    :func:`pandas.merge_asof` is designed to do this, but requires that dataframes are
-    sorted by the merge keys (``left_on``, ``right_on``, and ``by`` here). We also need
-    to make sure that all merge keys have identical data types in the two dataframes
-    (e.g. ``plant_id_eia`` needs to be a nullable integer in both dataframes, not a
-    python int in one, and a nullable :func:`pandas.Int64Dtype` in the other).  Note
-    that :func:`pandas.merge_asof` performs a left merge, so the higher frequency
-    dataframe **must** be the left dataframe.
-
-    We also force both ``left_on`` and ``right_on`` to be a Datetime using
-    :func:`pandas.to_datetime` to allow merging dataframes having integer years with
-    those having datetime columns.
-
-    Because :func:`pandas.merge_asof` searches backwards for the first matching date,
-    this function only works if the less granular dataframe uses the convention of
-    reporting the first date in the time period for which it reports. E.g. annual
-    dataframes need to have January 1st as the date. This is what happens by defualt if
-    only a year or year-month are provided to :func:`pandas.to_datetime` as strings.
+    If the column isn't a datetime, it needs to be converted to a string type
+    first so that integer years are formatted correctly.
 
     Args:
-        left: The higher frequency "data" dataframe. Typically monthly in our use
-            cases. E.g. ``generation_eia923``. Must contain ``report_date`` and any
-            columns specified in the ``by`` argument.
-        right: The lower frequency "attribute" dataframe. Typically annual in our uses
-            cases. E.g. ``generators_eia860``. Must contain ``report_date`` and any
-            columns specified in the ``by`` argument.
-        left_on: Column in ``left`` to merge on using merge_asof. Default is
-            ``report_date``. Must be convertible to a Datetime using
-            :func:`pandas.to_datetime`
-        right_on: Column in ``right`` to merge on using :func:`pd.merge_asof`.  Default
-            is ``report_date``. Must be convertible to a Datetime using
-            :func:`pandas.to_datetime`
-
-        by: Columns to merge on in addition to ``report_date``. Typically ID columns
-            like ``plant_id_eia``, ``generator_id`` or ``boiler_id``.
+        df (pandas.DataFrame): Dataframe with column to convert.
+        date_col_name (string): name of the column to convert.
 
     Returns:
-        Merged contents of left and right input dataframes.  Will be sorted by
-        ``left_on`` and any columns specified in ``by``. See documentation for
-        :func:`pandas.merge_asof` to understand how this kind of merge works.
+        Dataframe with the converted datetime column.
+    """
+    if pd.api.types.is_datetime64_ns_dtype(df[date_col_name]) is False:
+        logger.warning(
+            f"{date_col_name} is {df[date_col_name].dtype} column. Converting to datetime."
+        )
+        df[date_col_name] = pd.to_datetime(df[date_col_name].astype("string"))
+    return df
+
+
+def full_timeseries_date_merge(
+    left: pd.DataFrame,
+    right: pd.DataFrame,
+    on: list[str],
+    left_date_col: str = "report_date",
+    right_date_col: str = "report_date",
+    new_date_col: str = "report_date",
+    date_on: list[str] = ["year"],
+    how: Literal["inner", "outer", "left", "right", "cross"] = "inner",
+    report_at_start: bool = True,
+    freq: str = "MS",
+    **kwargs,
+):
+    """Merge dataframes with different date frequencies and expand to a full timeseries.
+
+    Arguments: see arguments for ``date_merge`` and ``expand_timeseries``
+    """
+    out = date_merge(
+        left=left,
+        right=right,
+        left_date_col=left_date_col,
+        right_date_col=right_date_col,
+        new_date_col=new_date_col,
+        on=on,
+        date_on=date_on,
+        how=how,
+        report_at_start=report_at_start,
+        **kwargs,
+    )
+    out = expand_timeseries(
+        df=out,
+        date_col=new_date_col,
+        freq=freq,
+        key_cols=on,
+    )
+    return out
+
+
+def _add_suffix_to_date_on(date_on):
+    """Check date_on list is valid and add _temp_for_merge suffix."""
+    if date_on is None:
+        date_on = ["year"]
+    date_on_suffix = []
+    for col in date_on:
+        if col not in ["year", "month", "quarter", "day"]:
+            raise AssertionError(
+                logger.error(f"{col} is not a valid string in date_on column list.")
+            )
+        date_on_suffix.append(col + "_temp_for_merge")
+    return date_on_suffix
+
+
+def date_merge(
+    left: pd.DataFrame,
+    right: pd.DataFrame,
+    on: list[str],
+    left_date_col: str = "report_date",
+    right_date_col: str = "report_date",
+    new_date_col: str = "report_date",
+    date_on: list[str] = None,
+    how: Literal["inner", "outer", "left", "right", "cross"] = "inner",
+    report_at_start: bool = True,
+    **kwargs,
+) -> pd.DataFrame:
+    """Merge two dataframes that have different report date frequencies.
+
+    We often need to bring together data that is reported at different
+    temporal granularities e.g. monthly basis versus annual basis. This function
+    acts as a wrapper on a pandas merge to allow merging at different temporal
+    granularities. The date columns of both dataframes are separated into
+    year, quarter, month, and day columns. Then, the dataframes are merged according
+    to ``how`` on the columns specified by the ``on`` and ``date_on`` argument,
+    which list the new temporal columns to merge on as well any additional shared columns.
+    Finally, the datetime column is reconstructed in the output dataframe and
+    named according to the ``new_date_col`` parameter.
+
+    Args:
+        left: The left dataframe in the merge. Typically monthly in our use
+            cases if doing a left merge E.g. ``generation_eia923``.
+            Must contain columns specified by ``left_date_col`` and
+            ``on`` argument.
+        right: The right dataframe in the merge. Typically annual in our uses
+            cases if doing a left merge E.g. ``generators_eia860``.
+            Must contain columns specified by ``right_date_col`` and ``on`` argument.
+        on: The columns to merge on that are shared between both
+            dataframes. Typically ID columns like ``plant_id_eia``, ``generator_id``
+            or ``boiler_id``.
+        left_date_col: Column in ``left`` containing datetime like data. Default is
+            ``report_date``. Must be a Datetime or convertible to a Datetime using
+            :func:`pandas.to_datetime`
+        right_date_col: Column in ``right`` containing datetime like data. Default is
+            ``report_date``. Must be a Datetime or convertible to a Datetime using
+            :func:`pandas.to_datetime`.
+        new_date_col: Name of the reconstructed datetime column in the output dataframe.
+        date_on: The temporal columns to merge on. Values in this list
+            of columns must be [``year``, ``quarter``, ``month``, ``day``].
+            E.g. if a monthly reported dataframe is being merged onto a daily reported
+            dataframe, then the merge would be performed on ``["year", "month"]``.
+            If one of these temporal columns already exists in the dataframe it will not
+            be clobbered by the merge, as the suffix "_temp_for_merge" is added when
+            expanding the datetime column into year, quarter, month, and day. By default,
+            `date_on` will just include year.
+        how: How the dataframes should be merged. See :func:`pandas.DataFrame.merge`.
+        report_at_start: Whether the data in the dataframe whose report date is not being
+            kept in the merged output (in most cases the less frequently reported dataframe)
+            is reported at the start or end of the time period e.g. January 1st
+            for annual data.
+        kwargs : Additional arguments to pass to :func:`pandas.DataFrame.merge`.
+
+
+    Returns:
+        Merged contents of left and right input dataframes.
 
     Raises:
-        ValueError: if ``left_on`` or ``right_on`` columns are missing from their
+        ValueError: if ``left_date_col`` or ``right_date_col`` columns are missing from their
             respective input dataframes.
-        ValueError: if any of the labels referenced in ``by`` are missing from either
+        ValueError: if any of the labels referenced in ``on`` are missing from either
             the left or right dataframes.
 
     """
-    # Make sure we've got all the required inputs...
-    if left_on not in left.columns:
-        raise ValueError(f"Left dataframe has no column {left_on}.")
-    if right_on not in right.columns:
-        raise ValueError(f"Right dataframe has no {right_on}.")
-    missing_left_cols = [col for col in by if col not in left.columns]
-    if missing_left_cols:
-        raise ValueError(f"Left dataframe is missing {missing_left_cols}.")
-    missing_right_cols = [col for col in by if col not in right.columns]
-    if missing_right_cols:
-        raise ValueError(f"Left dataframe is missing {missing_right_cols}.")
 
-    def cleanup(df, on, by):
-        df = apply_pudl_dtypes(df)
-        df.loc[:, on] = pd.to_datetime(df[on])
-        df = df.sort_values([on] + by)
+    def separate_date_cols(df, date_col_name, date_on):
+        df[date_col_name] = pd.to_datetime(df[date_col_name])
+        if "year_temp_for_merge" in date_on:
+            df.loc[:, "year_temp_for_merge"] = df[date_col_name].dt.year
+        if "quarter_temp_for_merge" in date_on:
+            df.loc[:, "quarter_temp_for_merge"] = df[date_col_name].dt.quarter
+        if "month_temp_for_merge" in date_on:
+            df.loc[:, "month_temp_for_merge"] = df[date_col_name].dt.month
+        if "day_temp_for_merge" in date_on:
+            df.loc[:, "day_temp_for_merge"] = df[date_col_name].dt.day
         return df
 
-    return pd.merge_asof(
-        cleanup(df=left, on=left_on, by=by),
-        cleanup(df=right, on=right_on, by=by),
-        left_on=left_on,
-        right_on=right_on,
-        by=by,
-        tolerance=pd.Timedelta("365 days"),  # Should never match across years.
+    right = convert_col_to_datetime(right, right_date_col)
+    left = convert_col_to_datetime(left, left_date_col)
+    date_on = _add_suffix_to_date_on(date_on)
+    right = separate_date_cols(right, right_date_col, date_on)
+    left = separate_date_cols(left, left_date_col, date_on)
+    merge_cols = date_on + on
+    out = pd.merge(left, right, on=merge_cols, how=how, **kwargs)
+
+    suffixes = ["", ""]
+    if left_date_col == right_date_col:
+        if "suffixes" in kwargs:
+            suffixes = kwargs["suffixes"]
+        else:
+            suffixes = ["_x", "_y"]
+    # reconstruct the new report date column and clean up columns
+    left_right_date_col = [left_date_col + suffixes[0], right_date_col + suffixes[1]]
+    if report_at_start:
+        # keep the later of the two report dates when determining
+        # the new report date for each row
+        reconstructed_date = out[left_right_date_col].max(axis=1)
+    else:
+        # keep the earlier of the two report dates
+        reconstructed_date = out[left_right_date_col].min(axis=1)
+    out = out.drop(left_right_date_col + date_on, axis=1)
+    out.insert(loc=0, column=new_date_col, value=reconstructed_date)
+    return out
+
+
+def expand_timeseries(
+    df: pd.DataFrame,
+    key_cols: list[str],
+    date_col: str = "report_date",
+    freq: str = "MS",
+    fill_through_freq: Literal["year", "month", "day"] = "year",
+) -> pd.DataFrame:
+    """Expand a dataframe to a include a full time series at a given frequency.
+
+    This function adds a full timeseries to the given dataframe for each group
+    of columns specified by ``key_cols``. The data in the timeseries will be filled
+    with the next previous chronological observation for a group of primary key columns
+    specified by ``key_cols``.
+
+    Arguments:
+        df: The dataframe to expand. Must have ``date_col`` in columns.
+        key_cols: Column names of the non-date primary key columns in the dataframe.
+            The resulting dataframe will have a full timeseries expanded for each
+            unique group of these ID columns that are present in the dataframe.
+        date_col: Name of the datetime column being expanded into a full timeseries.
+        freq: The frequency of the time series to expand the data to.
+            See :ref:`here <timeseries.offset_aliases>` for a list of
+            frequency aliases.
+        fill_through_freq: The frequency in which to fill in the data through. For
+            example, if equal to "year" the data will be filled in through the end of
+            the last reported year for each grouping of `key_cols`. Valid frequencies
+            are only year, month, or day.
+    """
+    try:
+        pd.tseries.frequencies.to_offset(freq)
+    except ValueError:
+        logger.exception(
+            f"Frequency string {freq} is not valid. \
+            See Pandas Timeseries Offset Aliases docs for valid strings."
+        )
+
+    # For each group of ID columns add a dummy record with the date column
+    # equal to one increment higher than the last record in the group for the
+    # desired fill_through_freq.
+    # This allows records to be filled through the end of the last reported period
+    # and then this dummy record is dropped
+    df = convert_col_to_datetime(df, date_col)
+    end_dates = df.groupby(key_cols).agg({date_col: "max"})
+    if fill_through_freq == "year":
+        end_dates.loc[:, date_col] = pd.to_datetime(
+            {
+                "year": end_dates[date_col].dt.year + 1,
+                "month": 1,
+                "day": 1,
+            }
+        )
+    elif fill_through_freq == "month":
+        end_dates.loc[:, date_col] = pd.to_datetime(
+            {
+                "year": end_dates[date_col].dt.year,
+                "month": end_dates[date_col].dt.month + 1,
+                "day": 1,
+            }
+        )
+    elif fill_through_freq == "day":
+        end_dates.loc[:, date_col] = pd.to_datetime(
+            {
+                "year": end_dates[date_col].dt.year,
+                "month": end_dates[date_col].dt.month,
+                "day": end_dates[date_col].dt.day + 1,
+            }
+        )
+    else:
+        raise AssertionError(
+            f"{fill_through_freq} is not a valid frequency to fill through."
+        )
+    end_dates["drop_row"] = True
+    df = pd.concat([df, end_dates.reset_index()])
+    df = (
+        df.set_index(date_col)
+        .groupby(key_cols)
+        .resample(freq)
+        .ffill()
+        .drop(key_cols, axis=1)
+        .reset_index()
     )
+    df = df[df.drop_row.isnull()].drop("drop_row", axis=1).reset_index(drop=True)
+    return df
 
 
 def organize_cols(df, cols):
@@ -447,7 +628,7 @@ def organize_cols(df, cols):
     """
     # Generate a list of all the columns in the dataframe that are not
     # included in cols
-    data_cols = sorted([c for c in df.columns.tolist() if c not in cols])
+    data_cols = sorted(c for c in df.columns.tolist() if c not in cols)
     organized_cols = cols + data_cols
     return df[organized_cols]
 
@@ -892,7 +1073,7 @@ def merge_dicts(list_of_dicts):
 
 
 def convert_cols_dtypes(
-    df: pd.DataFrame, data_source: Optional[str] = None, name: Optional[str] = None
+    df: pd.DataFrame, data_source: str | None = None, name: str | None = None
 ) -> pd.DataFrame:
     """Convert a PUDL dataframe's columns to the correct data type.
 
@@ -1308,7 +1489,7 @@ def sum_and_weighted_average_agg(
     df_in: pd.DataFrame,
     by: list,
     sum_cols: list,
-    wtavg_dict: Dict[str, str],
+    wtavg_dict: dict[str, str],
 ) -> pd.DataFrame:
     """Aggregate dataframe by summing and using weighted averages.
 
@@ -1365,7 +1546,7 @@ def get_eia_ferc_acct_map():
 
 def dedupe_n_flatten_list_of_lists(mega_list):
     """Flatten a list of lists and remove duplicates."""
-    return list(set([item for sublist in mega_list for item in sublist]))
+    return list({item for sublist in mega_list for item in sublist})
 
 
 def convert_df_to_excel_file(df: pd.DataFrame, **kwargs) -> pd.ExcelFile:
@@ -1386,7 +1567,7 @@ def convert_df_to_excel_file(df: pd.DataFrame, **kwargs) -> pd.ExcelFile:
     return pd.ExcelFile(workbook)
 
 
-def configure_root_logger(logfile: Optional[str] = None):
+def configure_root_logger(logfile: str | None = None):
     """Configure the root catalystcoop logger.
 
     Args:
